requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
psycopg2[binary]
validate-docbr